Rework Draco tests; add Draco_2CylinderEngine
The problem is that running Draco tests in the CI requires USD binaries
specifically built against Draco. Retro-deploying the last five USD
versions is a lot of work, so we only run a single test locally. On the
plus side, we now perform a graphical test which turns out to have a
slightly different ref image.